+++ tests/Export/ProductServiceTest.php
@@ -117,30 +117,77 @@ class ProductServiceTest extends TestCase
         $this->assertSame('FINDOLOGIC VARIANT', $product->getName());
     }
 
-    public function testIfMoreThanOneVariantExistsItWillUseVariantInformationInsteadOfMainProductInformation(): void
+    public function variantProvider(): array
     {
         $expectedParentId = Uuid::randomHex();
-        $expectedFirstVariantId = Uuid::randomHex();
-        $expectedSecondVariantId = Uuid::randomHex();
 
-        $this->createVisibleTestProduct([
-            'id' => $expectedParentId,
-            'active' => false
-        ]);
+        return [
+            'variant information is used instead of product information' => [
+                'mainProduct' => [
+                    'id' => $expectedParentId,
+                    'active' => false,
+                ],
+                'variants' => [
+                    [
+                        'id' => Uuid::randomHex(),
+                        'parentId' => $expectedParentId,
+                        'active' => true,
+                        'productNumber' => 'FINDOLOGIC001.1',
+                        'name' => 'FINDOLOGIC VARIANT 1',
+                    ],
+                    [
+                        'id' => Uuid::randomHex(),
+                        'parentId' => $expectedParentId,
+                        'active' => true,
+                        'productNumber' => 'FINDOLOGIC001.2',
+                        'name' => 'FINDOLOGIC VARIANT 2',
+                    ],
+                ],
+                'expectedChildCount' => 2,
+            ],
+            'inactive variants are not considered' => [
+                'mainProduct' => [
+                    'id' => $expectedParentId,
+                    'active' => true,
+                ],
+                'variants' => [
+                    [
+                        'id' => Uuid::randomHex(),
+                        'parentId' => $expectedParentId,
+                        'active' => false,
+                        'productNumber' => 'FINDOLOGIC001.1',
+                        'name' => 'FINDOLOGIC VARIANT 1',
+                    ],
+                    [
+                        'id' => Uuid::randomHex(),
+                        'parentId' => $expectedParentId,
+                        'active' => true,
+                        'productNumber' => 'FINDOLOGIC001.2',
+                        'name' => 'FINDOLOGIC VARIANT 2',
+                    ],
+                ],
+                'expectedChildCount' => 1,
+            ],
+        ];
+    }
 
-        $this->createVisibleTestProduct($this->getBasicVariantData([
-            'id' => $expectedFirstVariantId,
-            'parentId' => $expectedParentId,
-            'productNumber' => 'FINDOLOGIC001.1',
-            'name' => 'FINDOLOGIC VARIANT 1',
-        ]));
+    /**
+     * @dataProvider variantProvider
+     */
+    public function testVariantDataGetsExportedDependingOnActiveState(
+        array $mainProduct,
+        array $variants,
+        int $expectedChildCount
+    ): void {
+        $this->createVisibleTestProduct($mainProduct);
 
-        $this->createVisibleTestProduct($this->getBasicVariantData([
-            'id' => $expectedSecondVariantId,
-            'parentId' => $expectedParentId,
-            'productNumber' => 'FINDOLOGIC001.2',
-            'name' => 'FINDOLOGIC VARIANT 2',
-        ]));
+        $variantIds = [];
+        foreach ($variants as $variant) {
+            $variantIds[] = $variant['id'];
+            $this->createVisibleTestProduct(
+                $this->getBasicVariantData($variant)
+            );
+        }
 
         $products = $this->defaultProductService->searchVisibleProducts(20, 0);
         /** @var ProductEntity $product */
@@ -150,6 +197,8 @@ class ProductServiceTest extends TestCase
         // sometimes the second statement may be executed before the first one, which causes a different result.
         // To prevent this test from failing if Shopware decides to create the second variant before the first one,
         // we ensure that the second variant is used instead.
+        $expectedFirstVariantId = $variantIds[0];
+        $expectedSecondVariantId = $variantIds[1];
         $expectedChildVariantId = $expectedSecondVariantId;
         try {
             $this->assertSame($expectedFirstVariantId, $product->getId());
@@ -158,11 +207,11 @@ class ProductServiceTest extends TestCase
             $expectedChildVariantId = $expectedFirstVariantId;
         }
 
-        $this->assertCount(2, $product->getChildren());
+        $this->assertCount($expectedChildCount, $product->getChildren());
 
         foreach ($product->getChildren() as $child) {
             if (!$child->getParentId()) {
-                $this->assertSame($expectedParentId, $child->getId());
+                $this->assertSame($mainProduct['id'], $child->getId());
             } else {
                 $this->assertSame($expectedChildVariantId, $child->getId());
             }